# docs/swagger.yaml
definitions:
...
  controllers.PostsAndUserDetail:
    properties:
      posts:
        items:
          $ref: '#/definitions/models.PostDetail'
        type: array
      user:
        $ref: '#/definitions/models.UserDetail'
    type: object
...
  models.ExtendedFile:
    properties:
      bucket:
        type: string
      create_at:
        type: string
      file_id:
        type: integer
      filename:
        type: string
      postID:
        type: integer
    type: object
...
  models.Hole:
    properties:
      content:
        type: string
      create_at:
        type: string
      forum_id:
        type: integer
      hole_id:
        type: integer
      title:
        type: string
      user_id:
        type: integer
    type: object
  models.HoleDetail:
    properties:
      content:
        type: string
      create_at:
        type: string
      forum_id:
        type: integer
      hole_id:
        type: integer
      title:
        type: string
      user_id:
        type: integer
    type: object
  models.PostDetail:
    properties:
      content:
        type: string
      create_at:
        type: string
      files:
        items:
          $ref: '#/definitions/models.ExtendedFile'
        type: array
      forum_id:
        type: integer
      like:
        type: integer
      post_id:
        type: integer
      title:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  models.UserDetail:
    properties:
      avatar:
        type: string
      create_at:
        type: string
      email:
        type: string
      is_admin:
        type: boolean
      like_list:
        items:
          type: integer
        type: array
      participate_list:
        items:
          type: integer
        type: array
      password:
        type: string
      star_list:
        items:
          type: integer
        type: array
      user_id:
        type: integer
      username:
        type: string
    type: object
  responses.StatusBadRequestResponse:
    properties:
      code:
        example: 400
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
...
  responses.StatusInternalServerError:
    properties:
      code:
        example: 500
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  responses.StatusOKResponse:
    properties:
      code:
        example: 200
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /forums/{forum_id}/holes:
    get:
      consumes:
      - application/json
      description: GetAllHolesByForumID
      parameters:
      - description: 将token放在请求头部的‘Authorization‘字段中，并以‘Bearer ‘开头
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取树洞帖子成功
          schema:
            allOf:
            - $ref: '#/definitions/responses.StatusOKResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Hole'
                  type: array
              type: object
        "500":
          description: 查询数据库出现异常
          schema:
            $ref: '#/definitions/responses.StatusInternalServerError'
      summary: GetAllHolesByForumID
      tags:
      - Holes
    post:
      consumes:
      - multipart/form-data
      description: CreateHole
      parameters:
      - description: 将token放在请求头部的‘Authorization‘字段中，并以‘Bearer ‘开头
        in: header
        name: token
        required: true
        type: string
      - description: Hole 的标题
        in: formData
        name: title
        required: true
        type: string
      - description: Hole 的内容
        in: formData
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 创建 hole 成功
          schema:
            $ref: '#/definitions/responses.StatusOKResponse'
        "400":
          description: 树洞的标题或者内容不得为空
          schema:
            $ref: '#/definitions/responses.StatusBadRequestResponse'
        "500":
          description: 插入用户创建的hole失败
          schema:
            $ref: '#/definitions/responses.StatusInternalServerError'
      summary: CreateHole
      tags:
      - Holes
  /forums/{forum_id}/holes/{hole_id}:
    get:
      consumes:
      - application/json
      description: GetOneHoleDetailByHoleID
      parameters:
      - description: 将token放在请求头部的‘Authorization‘字段中，并以‘Bearer ‘开头
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取树洞帖子成功
          schema:
            allOf:
            - $ref: '#/definitions/responses.StatusOKResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.HoleDetail'
                  type: array
              type: object
        "400":
          description: 数据库查询异常，或者该hole不存在
          schema:
            $ref: '#/definitions/responses.StatusInternalServerError'
      summary: GetOneHoleDetailByHoleID
      tags:
      - Holes
  /forums/{forum_id}/posts:
    get:
      consumes:
      - application/json
      description: GetAllPostsByForumID
      parameters:
      - description: 将token放在请求头部的‘Authorization‘字段中，并以‘Bearer ‘开头
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.StatusOKResponse'
            - properties:
                data:
                  $ref: '#/definitions/controllers.PostsAndUserDetail'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/responses.StatusInternalServerError'
      summary: GetAllPostsByForumID
      tags:
      - Posts
    post:
      consumes:
      - multipart/form-data
      description: CreatePost
      parameters:
      - description: 将token放在请求头部的‘Authorization‘字段中，并以‘Bearer ‘开头
        in: header
        name: token
        required: true
        type: string
      - description: Post 的标题
        in: formData
        name: title
        required: true
        type: string
      - description: Post 的内容
        in: formData
        name: content
        required: true
        type: string
      - description: 文件内容
        in: formData
        name: files[]
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 创建 Post 成功
          schema:
            $ref: '#/definitions/responses.StatusOKResponse'
        "403":
          description: 您所上传的文件无法打开
          schema:
            $ref: '#/definitions/responses.StatusBadRequestResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/responses.StatusInternalServerError'
      summary: CreatePost
      tags:
      - Posts
  /forums/{forum_id}/posts/{post_id}:
    get:
      consumes:
      - application/json
      description: GetOnePostDetailByPostID
      parameters:
      - description: 将token放在请求头部的‘Authorization‘字段中，并以‘Bearer ‘开头
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.StatusOKResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.PostDetail'
                  type: array
              type: object
        "400":
          description: 数据库查询异常，或者该post不存在
          schema:
            $ref: '#/definitions/responses.StatusInternalServerError'
      summary: GetOnePostDetailByPostID
      tags:
      - Posts
...
swagger: "2.0"